Refactor command line interface
Separate into runner functions used in the unit tests,
and command line tools

Use the Annotated[..] for type safety
