+++ src/subhkl/commands.py
@@ -1,0 +1,765 @@
+# src/subhkl/io/commands.py
+import glob
+import h5py
+import numpy as np
+
+# NOTE(Vivek): deprecate and use Goniometer class to handler rotation calc
+from subhkl.instrument.goniometer import (
+    calc_goniometer_rotation_matrix,
+    get_rotation_data_from_nexus,
+)
+from subhkl.io.export import FinderConcatenateMerger, ImageStackMerger, MTZExporter
+from subhkl.integration import Peaks
+from subhkl.instrument.metrics import compute_metrics
+from subhkl.optimization import FindUB
+
+
+def run_index(
+    hdf5_peaks_filename: str | None = None,
+    output_peaks_filename: str | None = None,
+    strategy_name: str = "DE",
+    population_size: int = 1000,
+    gens: int = 100,
+    n_runs: int = 1,
+    seed: int = 0,
+    tolerance_deg: float = 0.1,
+    sigma_init: float | None = None,
+    refine_lattice: bool = False,
+    lattice_bound_frac: float = 0.05,
+    bootstrap_filename: str | None = None,
+    refine_goniometer: bool = False,
+    refine_goniometer_axes: list | None = None,
+    goniometer_bound_deg: float = 5.0,
+    refine_sample: bool = False,
+    sample_bound_meters: float = 0.002,
+    refine_beam: bool = False,
+    beam_bound_deg: float = 1.0,
+    nexus_filename: str | None = None,
+    instrument_name: str | None = None,
+    loss_method: str = "cosine",
+    hkl_search_range: int = 20,
+    d_min: float | None = None,
+    d_max: float | None = None,
+    search_window_size: int = 512,
+    batch_size: int | None = None,
+    window_batch_size: int = 32,
+    chunk_size: int = 256,
+    num_iters: int = 20,
+    top_k: int = 32,
+    B_sharpen: float | None = None,
+    input_data: dict | None = None,
+    wavelength_min: float | None = None,
+    wavelength_max: float | None = None,
+):
+    """
+    Index the given peak file and save it using the evosax optimizer.
+    """
+    if input_data is not None:
+        opt = FindUB(data=input_data)
+    else:
+        opt = FindUB(filename=hdf5_peaks_filename)
+
+    if wavelength_min is not None and wavelength_max is not None:
+        opt.wavelength = [wavelength_min, wavelength_max]
+
+    print(f"Starting evosax optimization with strategy: {strategy_name}")
+    print(f"Running {n_runs} run(s)...")
+    print(f"Settings per run: Population Size={population_size}, Generations={gens}")
+    if refine_lattice:
+        print(f"Refining lattice parameters with {lattice_bound_frac * 100}% bounds.")
+    if refine_sample:
+        print(f"Refining sample offset with {1000 * sample_bound_meters} mm bounds.")
+    if refine_beam:
+        print(f"Refining beam tilt with {beam_bound_deg}° bounds.")
+
+    goniometer_names = None
+    if refine_goniometer:
+        if nexus_filename and instrument_name:
+            print(
+                f"Refining goniometer angles from Nexus with {goniometer_bound_deg} deg bounds."
+            )
+            axes, angles, names = get_rotation_data_from_nexus(
+                nexus_filename, instrument_name
+            )
+            opt.goniometer_axes = np.array(axes)
+
+            if opt.run_indices is not None:
+                num_runs = np.max(opt.run_indices) + 1
+                opt.goniometer_angles = np.array(angles)[:, np.newaxis].repeat(
+                    num_runs, axis=1
+                )
+            else:
+                num_peaks = len(opt.two_theta)
+                opt.goniometer_angles = np.array(angles)[:, np.newaxis].repeat(
+                    num_peaks, axis=1
+                )
+            goniometer_names = names
+        elif opt.goniometer_axes is not None:
+            print(
+                f"Refining goniometer angles from HDF5 file with {goniometer_bound_deg} deg bounds."
+            )
+        else:
+            print(
+                "WARNING: refine_goniometer requested but goniometer data not found. Skipping."
+            )
+            refine_goniometer = False
+
+    init_params = None
+    if bootstrap_filename:
+        init_params = opt.get_bootstrap_params(
+            bootstrap_filename,
+            refine_lattice=refine_lattice,
+            lattice_bound_frac=lattice_bound_frac,
+            refine_sample=refine_sample,
+            sample_bound_meters=sample_bound_meters,
+            refine_beam=refine_beam,
+            beam_bound_deg=beam_bound_deg,
+            refine_goniometer=refine_goniometer,
+            goniometer_bound_deg=goniometer_bound_deg,
+            refine_goniometer_axes=refine_goniometer_axes,
+        )
+
+    num, hkl, lamda, U = opt.minimize(
+        strategy_name=strategy_name,
+        population_size=population_size,
+        num_generations=gens,
+        n_runs=n_runs,
+        sigma_init=sigma_init,
+        seed=seed,
+        tolerance_deg=tolerance_deg,
+        init_params=init_params,
+        refine_lattice=refine_lattice,
+        lattice_bound_frac=lattice_bound_frac,
+        refine_goniometer=refine_goniometer,
+        refine_goniometer_axes=refine_goniometer_axes,
+        goniometer_bound_deg=goniometer_bound_deg,
+        goniometer_names=goniometer_names,
+        refine_sample=refine_sample,
+        sample_bound_meters=sample_bound_meters,
+        refine_beam=refine_beam,
+        beam_bound_deg=beam_bound_deg,
+        loss_method=loss_method,
+        d_min=d_min,
+        d_max=d_max,
+        hkl_search_range=hkl_search_range,
+        search_window_size=search_window_size,
+        batch_size=batch_size,
+        window_batch_size=window_batch_size,
+        chunk_size=chunk_size,
+        num_iters=num_iters,
+        top_k=top_k,
+        B_sharpen=B_sharpen,
+    )
+
+    print(f"\nOptimization complete. Best solution indexed {num} peaks.")
+
+    opt.reciprocal_lattice_B()
+
+    copy_keys = [
+        "sample/space_group",
+        "instrument/wavelength",
+        "peaks/intensity",
+        "peaks/sigma",
+        "peaks/two_theta",
+        "peaks/azimuthal",
+        "peaks/radius",
+        "peaks/xyz",
+        "goniometer/R",
+        "goniometer/axes",
+        "goniometer/angles",
+        "goniometer/names",
+        "files",
+        "file_offsets",
+        "peaks/run_index",
+        "bank",
+        "sample/offset",
+        "beam/ki_vec",
+    ]
+
+    copied_data = {}
+
+    if input_data is not None:
+        for key in copy_keys:
+            if key in input_data:
+                copied_data[key] = input_data[key]
+    else:
+        with h5py.File(hdf5_peaks_filename, "r") as f:
+            for key in copy_keys:
+                if key in f:
+                    copied_data[key] = np.array(f[key])
+
+    print(f"Saving indexed peaks to {output_peaks_filename}...")
+    with h5py.File(output_peaks_filename, "w") as f:
+        if instrument_name:
+            f.attrs["instrument"] = instrument_name
+        elif "instrument" in input_data:
+            f.attrs["instrument"] = input_data["instrument"]
+        for key, value in copied_data.items():
+            f[key] = value
+
+        def safe_write(grp, name, data):
+            if name in grp:
+                del grp[name]
+            grp[name] = data
+
+        safe_write(f, "goniometer/R", opt.R)
+
+        if opt.goniometer_offsets is not None:
+            safe_write(f, "optimization/goniometer_offsets", opt.goniometer_offsets)
+
+        if opt.sample_offset is not None:
+            safe_write(f, "sample/offset", opt.sample_offset)
+
+        if opt.ki_vec is not None:
+            safe_write(f, "beam/ki_vec", opt.ki_vec)
+
+        safe_write(f, "sample/a", opt.a)
+        safe_write(f, "sample/b", opt.b)
+        safe_write(f, "sample/c", opt.c)
+        safe_write(f, "sample/alpha", opt.alpha)
+        safe_write(f, "sample/beta", opt.beta)
+        safe_write(f, "sample/gamma", opt.gamma)
+
+        B_mat = opt.reciprocal_lattice_B()
+        safe_write(f, "sample/B", B_mat)
+        f["sample/U"] = U
+
+        if opt.run_indices is not None:
+            safe_write(f, "peaks/run_index", opt.run_indices)
+
+        f["peaks/h"] = hkl[:, 0]
+        f["peaks/k"] = hkl[:, 1]
+        f["peaks/l"] = hkl[:, 2]
+        f["peaks/lambda"] = lamda
+        f["optimization/best_params"] = opt.x
+    print("Done.")
+
+
+def run_finder(
+    filename: str,
+    instrument: str,
+    output_filename: str = "output.h5",
+    finder_algorithm: str = "peak_local_max",
+    show_progress: bool = True,
+    create_visualizations: bool = False,
+    show_steps: bool = False,
+    peak_local_max_min_pixel_distance: int = -1,
+    peak_local_max_min_relative_intensity: float = -1,
+    peak_local_max_normalization: bool = False,
+    thresholding_noise_cutoff_quantile: float = 0.8,
+    thresholding_min_peak_dist_pixels: float = 8.0,
+    thresholding_mask_file: str | None = None,
+    thresholding_mask_rel_erosion_radius: float = 0.05,
+    thresholding_blur_kernel_sigma: int = 5,
+    thresholding_open_kernel_size_pixels: int = 3,
+    wavelength_min: float | None = None,
+    wavelength_max: float | None = None,
+    region_growth_distance_threshold: float = 1.5,
+    region_growth_minimum_sigma: float | None = None,
+    region_growth_minimum_intensity: float = 4500.0,
+    region_growth_maximum_pixel_radius: float = 17.0,
+    peak_center_box_size: int = 15,
+    peak_smoothing_window_size: int = 15,
+    peak_minimum_pixels: int = 30,
+    peak_minimum_signal_to_noise: float = 1.0,
+    peak_pixel_outlier_threshold: float = 2.0,
+    sparse_rbf_alpha: float = 0.1,
+    sparse_rbf_gamma: float = 2.0,
+    sparse_rbf_min_sigma: float = 0.5,
+    sparse_rbf_max_sigma: float = 10.0,
+    sparse_rbf_max_peaks: int = 500,
+    sparse_rbf_chunk_size: int = 4096,
+    sparse_rbf_tile_rows: int = 2,
+    sparse_rbf_tile_cols: int = 2,
+    max_workers: int = 16,
+):
+    print(f"Creating peaks from {filename} for instrument {instrument}")
+
+    wavelength_kwargs = {}
+    if wavelength_min:
+        wavelength_kwargs["wavelength_min"] = wavelength_min
+    if wavelength_max:
+        wavelength_kwargs["wavelength_max"] = wavelength_max
+
+    peaks = Peaks(filename, instrument, **wavelength_kwargs)
+
+    peak_kwargs = {"algorithm": finder_algorithm}
+    if finder_algorithm == "peak_local_max":
+        if peak_local_max_min_pixel_distance > 0:
+            peak_kwargs["min_pix"] = peak_local_max_min_pixel_distance
+        if peak_local_max_min_relative_intensity > 0:
+            peak_kwargs["min_rel_intensity"] = peak_local_max_min_relative_intensity
+        peak_kwargs["normalize"] = peak_local_max_normalization
+    elif finder_algorithm == "thresholding":
+        peak_kwargs.update(
+            {
+                "noise_cutoff_quantile": thresholding_noise_cutoff_quantile,
+                "min_peak_dist_pixels": thresholding_min_peak_dist_pixels,
+                "mask_file": thresholding_mask_file,
+                "mask_rel_erosion_radius": thresholding_mask_rel_erosion_radius,
+                "blur_kernel_sigma": thresholding_blur_kernel_sigma,
+                "open_kernel_size_pixels": thresholding_open_kernel_size_pixels,
+                "show_steps": show_steps,
+                "show_scale": "log",
+            }
+        )
+    elif finder_algorithm == "sparse_rbf":
+        peak_kwargs.update(
+            {
+                "alpha": sparse_rbf_alpha,
+                "gamma": sparse_rbf_gamma,
+                "min_sigma": sparse_rbf_min_sigma,
+                "max_sigma": sparse_rbf_max_sigma,
+                "max_peaks": sparse_rbf_max_peaks,
+                "chunk_size": sparse_rbf_chunk_size,
+                "show_steps": show_steps,
+                "show_scale": "linear",
+                "tiles": (sparse_rbf_tile_rows, sparse_rbf_tile_cols),
+            }
+        )
+    else:
+        raise ValueError("Invalid finder algorithm")
+
+    integration_params = {
+        "region_growth_distance_threshold": region_growth_distance_threshold,
+        "region_growth_minimum_sigma": region_growth_minimum_sigma,
+        "region_growth_minimum_intensity": region_growth_minimum_intensity,
+        "region_growth_maximum_pixel_radius": region_growth_maximum_pixel_radius,
+        "peak_center_box_size": peak_center_box_size,
+        "peak_smoothing_window_size": peak_smoothing_window_size,
+        "peak_minimum_pixels": peak_minimum_pixels,
+        "peak_minimum_signal_to_noise": peak_minimum_signal_to_noise,
+        "peak_pixel_outlier_threshold": peak_pixel_outlier_threshold,
+    }
+
+    detector_peaks = peaks.get_detector_peaks(
+        peak_kwargs,
+        integration_params,
+        visualize=create_visualizations,
+        show_progress=show_progress,
+        file_prefix=filename,
+        max_workers=max_workers,
+    )
+
+    peaks.write_hdf5(
+        output_filename=output_filename,
+        detector_peaks=detector_peaks,
+        instrument_wavelength=[peaks.wavelength.min, peaks.wavelength.max],
+    )
+
+
+def run_finder_merger(
+    finder_h5_txt_list_filename: str,
+    output_pre_index_filename: str,
+    a: float,
+    b: float,
+    c: float,
+    alpha: float,
+    beta: float,
+    gamma: float,
+    wavelength_min: float,
+    wavelength_max: float,
+    space_group: str,
+):
+    with open(finder_h5_txt_list_filename) as f:
+        finder_h5_files = f.read().splitlines()
+
+    merging_algorithm = FinderConcatenateMerger(finder_h5_files)
+    merging_algorithm.merge(output_pre_index_filename)
+
+    with h5py.File(output_pre_index_filename, "r+") as f:
+        f["sample/a"] = a
+        f["sample/b"] = b
+        f["sample/c"] = c
+        f["sample/alpha"] = alpha
+        f["sample/beta"] = beta
+        f["sample/gamma"] = gamma
+        f["sample/space_group"] = space_group
+        f["instrument/wavelength"] = [wavelength_min, wavelength_max]
+
+
+def run_metrics(
+    filename: str,
+    found_peaks_file: str | None = None,
+    instrument: str | None = None,
+    d_min: float | None = None,
+    per_run: bool = False,
+):
+    result = compute_metrics(
+        filename=filename,
+        found_peaks_file=found_peaks_file,
+        instrument=instrument,
+        d_min=d_min,
+        per_run=per_run,
+    )
+
+    if "error_message" in result:
+        print(result["error_message"])
+        if result["error_message"].startswith("Exception"):
+            print("METRICS: 9.99 9.99 9.99 9.99 9.99 9.99")
+        return
+
+    if "filter_message" in result:
+        print(f"METRICS: {result['filter_message']}")
+
+    print(
+        f"METRICS: {result['median_d_err']:.5f} {result['mean_d_err']:.5f} {result['max_d_err']:.5f} "
+        f"{result['median_ang_err']:.5f} {result['mean_ang_err']:.5f} {result['max_ang_err']:.5f}"
+    )
+
+    if per_run and "per_run_errors" in result:
+        print("\nPER-RUN MEDIAN ANGULAR ERROR (deg) - Sorted by error:")
+        for r, err, count in result["per_run_errors"]:
+            status = "BAD" if err > 1.0 else "OK"
+            print(f"  Run {r:4d}: {err:6.3f} ({count:4d} peaks) [{status}]")
+
+
+def run_peak_predictor(
+    filename: str,
+    instrument: str,
+    indexed_hdf5_filename: str,
+    integration_peaks_filename: str,
+    d_min: float = 1.0,
+    create_visualizations: bool = False,
+    space_group: str | None = None,
+    wavel_min: float | None = None,
+    wavel_max: float | None = None,
+    max_workers: int = 16,
+):
+    with h5py.File(indexed_hdf5_filename, "r") as f_idx:
+        a = float(f_idx["sample/a"][()])
+        b = float(f_idx["sample/b"][()])
+        c = float(f_idx["sample/c"][()])
+        alpha = float(f_idx["sample/alpha"][()])
+        beta = float(f_idx["sample/beta"][()])
+        gamma = float(f_idx["sample/gamma"][()])
+
+        if space_group is None:
+            space_group = f_idx["sample/space_group"][()].decode("utf-8")
+
+        wavelength = f_idx["instrument/wavelength"][()]
+        if wavel_min:
+            wavelength[0] = wavel_min
+        if wavel_max:
+            wavelength[1] = wavel_max
+
+        U = f_idx["sample/U"][()]
+        B = f_idx["sample/B"][()]
+
+        offsets = (
+            f_idx["optimization/goniometer_offsets"][()]
+            if "optimization/goniometer_offsets" in f_idx
+            else None
+        )
+        sample_offset = (
+            f_idx["sample/offset"][()] if "sample/offset" in f_idx else np.zeros(3)
+        )
+        ki_vec = (
+            f_idx["beam/ki_vec"][()]
+            if "beam/ki_vec" in f_idx
+            else np.array([0.0, 0.0, 1.0])
+        )
+
+    peaks = Peaks(
+        filename, instrument, wavelength_min=wavelength[0], wavelength_max=wavelength[1]
+    )
+    print(
+        f"Predicting peaks for {len(peaks.image.ims)} images using solution from {indexed_hdf5_filename}"
+    )
+
+    all_R = peaks.goniometer.rotation
+
+    if offsets is not None:
+        print(f"Applying refined goniometer offsets from indexer: {offsets}")
+        if (
+            peaks.goniometer.angles_raw is not None
+            and peaks.goniometer.axes_raw is not None
+        ):
+            angles_refined = peaks.goniometer.angles_raw + offsets[None, :]
+            all_R = np.stack(
+                [
+                    calc_goniometer_rotation_matrix(peaks.goniometer.axes_raw, ang)
+                    for ang in angles_refined
+                ]
+            )
+        else:
+            print("WARNING: Cannot apply refined offsets. Using nominal R stack.")
+    else:
+        print("Using nominal R stack directly from raw images (no offsets applied).")
+
+    UB = U @ B
+    if all_R.ndim == 3:
+        RUB = np.matmul(all_R, UB)
+    else:
+        RUB = all_R @ UB
+
+    results_map = peaks.predict_peaks(
+        a,
+        b,
+        c,
+        alpha,
+        beta,
+        gamma,
+        d_min,
+        RUB=RUB,
+        space_group=space_group,
+        sample_offset=sample_offset,
+        ki_vec=ki_vec,
+        max_workers=max_workers,
+        R_all=all_R,
+    )
+
+    print(f"Saving predictions to {integration_peaks_filename}")
+    with h5py.File(integration_peaks_filename, "w") as f:
+        f.attrs["instrument"] = instrument
+        f["sample/a"], f["sample/b"], f["sample/c"] = a, b, c
+        f["sample/alpha"], f["sample/beta"], f["sample/gamma"] = alpha, beta, gamma
+
+        sorted_keys = sorted(peaks.image.ims.keys())
+        bank_ids = np.array(
+            [peaks.image.bank_mapping.get(k, k) for k in sorted_keys], dtype=np.int32
+        )
+        f.create_dataset("bank_ids", data=bank_ids)
+
+        f["sample/space_group"] = space_group
+        f["sample/U"], f["sample/B"] = U, B
+        f["instrument/wavelength"] = wavelength
+        f["goniometer/R"] = all_R
+
+        try:
+            goniometer_angles_to_save = angles_refined
+        except NameError:
+            goniometer_angles_to_save = peaks.goniometer.angles_raw
+
+        f["goniometer/angles"] = goniometer_angles_to_save
+        f["goniometer/axes"] = peaks.goniometer.axes_raw
+        if peaks.goniometer.names_raw:
+            dt = h5py.string_dtype(encoding="utf-8")
+            f.create_dataset(
+                "goniometer/names", data=peaks.goniometer.names_raw, dtype=dt
+            )
+
+        f["sample/offset"] = sample_offset
+        f["beam/ki_vec"] = ki_vec
+
+        for img_key, (i, j, h, k, l, wl) in results_map.items():
+            grp = f.create_group(f"banks/{img_key}")
+            grp.create_dataset("i", data=i)
+            grp.create_dataset("j", data=j)
+            grp.create_dataset("h", data=h)
+            grp.create_dataset("k", data=k)
+            grp.create_dataset("l", data=l)
+            grp.create_dataset("wavelength", data=wl)
+
+
+def run_integrator(
+    filename: str,
+    instrument: str,
+    integration_peaks_filename: str,
+    output_filename: str,
+    integration_method: str = "free_fit",
+    integration_mask_file: str | None = None,
+    integration_mask_rel_erosion_radius: float | None = 0.05,
+    region_growth_distance_threshold: float = 1.5,
+    region_growth_minimum_intensity: float = 50.0,
+    region_growth_minimum_sigma: float | None = None,
+    region_growth_maximum_pixel_radius: float = 17.0,
+    peak_center_box_size: int = 15,
+    peak_smoothing_window_size: int = 15,
+    peak_minimum_pixels: int = 10,
+    peak_minimum_signal_to_noise: float = 1.0,
+    peak_pixel_outlier_threshold: float = 2.0,
+    create_visualizations: bool = False,
+    show_progress: bool = True,
+    found_peaks_file: str | None = None,
+    max_workers: int = 16,
+):
+    peak_dict = {}
+    angles_stack = None
+    all_R = None
+    with h5py.File(integration_peaks_filename, "r") as f:
+        U = f["sample/U"][()] if "sample/U" in f else None
+        B = f["sample/B"][()] if "sample/B" in f else None
+        all_R = f["goniometer/R"][()] if "goniometer/R" in f else None
+        angles_stack = f["goniometer/angles"][()] if "goniometer/angles" in f else None
+        sample_offset = f["sample/offset"][()] if "sample/offset" in f else np.zeros(3)
+        ki_vec = (
+            f["beam/ki_vec"][()] if "beam/ki_vec" in f else np.array([0.0, 0.0, 1.0])
+        )
+
+        for key in f["banks"].keys():
+            img_idx = int(key)
+            grp = f[f"banks/{key}"]
+            peak_dict[img_idx] = [
+                grp["i"][()],
+                grp["j"][()],
+                grp["h"][()],
+                grp["k"][()],
+                grp["l"][()],
+                grp["wavelength"][()],
+            ]
+
+    peaks = Peaks(filename, instrument)
+
+    integration_params = {
+        "region_growth_distance_threshold": region_growth_distance_threshold,
+        "region_growth_minimum_intensity": region_growth_minimum_intensity,
+        "region_growth_minimum_sigma": region_growth_minimum_sigma,
+        "region_growth_maximum_pixel_radius": region_growth_maximum_pixel_radius,
+        "peak_center_box_size": peak_center_box_size,
+        "peak_smoothing_window_size": peak_smoothing_window_size,
+        "peak_minimum_pixels": peak_minimum_pixels,
+        "peak_minimum_signal_to_noise": peak_minimum_signal_to_noise,
+        "peak_pixel_outlier_threshold": peak_pixel_outlier_threshold,
+        "integration_mask_file": integration_mask_file,
+        "integration_mask_rel_erosion_radius": integration_mask_rel_erosion_radius,
+    }
+
+    if all_R is None:
+        print("Warning: Refined R stack not found in prediction file. Using nominal.")
+        all_R = peaks.goniometer.rotation
+
+    if angles_stack is None:
+        angles_stack = peaks.goniometer.angles_raw
+
+    UB = U @ B if U is not None and B is not None else None
+    RUB = None
+    if UB is not None:
+        RUB = np.matmul(all_R, UB) if all_R.ndim == 3 else all_R @ UB
+
+    result = peaks.integrate(
+        peak_dict,
+        integration_params,
+        RUB=RUB,
+        R_stack=all_R,
+        angles_stack=angles_stack,
+        sample_offset=sample_offset,
+        ki_vec=ki_vec,
+        create_visualizations=create_visualizations,
+        show_progress=show_progress,
+        integration_method=integration_method,
+        file_prefix=filename,
+        found_peaks_file=found_peaks_file,
+        max_workers=max_workers,
+    )
+
+    print(f"Saving integrated peaks to {output_filename}")
+
+    copy_keys = [
+        "sample/a",
+        "sample/b",
+        "sample/c",
+        "sample/alpha",
+        "sample/beta",
+        "sample/gamma",
+        "sample/space_group",
+        "sample/U",
+        "sample/B",
+        "sample/offset",
+        "beam/ki_vec",
+        "instrument/wavelength",
+    ]
+
+    with h5py.File(output_filename, "w") as f:
+        f["peaks/h"], f["peaks/k"], f["peaks/l"] = result.h, result.k, result.l
+        f["peaks/lambda"] = result.wavelength
+        f["peaks/intensity"], f["peaks/sigma"] = result.intensity, result.sigma
+        f["peaks/two_theta"], f["peaks/azimuthal"] = result.tt, result.az
+        f["peaks/bank"] = result.bank
+        f["peaks/run_index"] = result.run_id
+        f["peaks/xyz"] = result.xyz
+
+        if result.R and any(r is not None for r in result.R):
+            f["goniometer/R"] = np.array(result.R)
+        if result.angles and any(a is not None for a in result.angles):
+            f["goniometer/angles"] = np.array(result.angles)
+
+        with h5py.File(integration_peaks_filename, "r") as f_in:
+            for key in copy_keys:
+                if key in f_in:
+                    f_in.copy(f_in[key], f, key)
+            for k in ["goniometer/axes", "goniometer/names"]:
+                if k in f_in:
+                    f_in.copy(f_in[k], f, k)
+
+
+def run_mtz_exporter(
+    indexed_h5_filename: str, output_mtz_filename: str, space_group: str
+):
+    algorithm = MTZExporter(indexed_h5_filename, space_group)
+    algorithm.write_mtz(output_mtz_filename)
+
+
+def run_reduce(
+    nexus_filename: str,
+    output_filename: str,
+    instrument: str,
+    wavelength_min: float | None = None,
+    wavelength_max: float | None = None,
+):
+    print(f"Reducing {nexus_filename} -> {output_filename}")
+    peaks_handler = Peaks(
+        nexus_filename,
+        instrument,
+        wavelength_min=wavelength_min,
+        wavelength_max=wavelength_max,
+    )
+
+    if not peaks_handler.image.ims:
+        print("Warning: No images found in file.")
+        return
+
+    sorted_banks = sorted(peaks_handler.image.ims.keys())
+    image_stack = np.stack([peaks_handler.image.ims[b] for b in sorted_banks])
+    bank_ids = np.array(sorted_banks, dtype=np.int32)
+    n_images = len(sorted_banks)
+
+    if peaks_handler.goniometer.angles_raw is not None:
+        angles_repeated = np.tile(peaks_handler.goniometer.angles_raw, (n_images, 1))
+    else:
+        angles_repeated = np.zeros((n_images, 3))
+
+    axes = (
+        np.array(peaks_handler.goniometer.axes_raw)
+        if peaks_handler.goniometer.axes_raw is not None
+        else np.array([0.0, 1.0, 0.0])
+    )
+
+    with h5py.File(output_filename, "w") as f:
+        f.create_dataset("images", data=image_stack, compression="lzf")
+        f.create_dataset("bank_ids", data=bank_ids)
+        f.create_dataset("goniometer/angles", data=angles_repeated)
+        f.create_dataset("goniometer/axes", data=axes)
+
+        if peaks_handler.goniometer.names_raw:
+            dt = h5py.string_dtype(encoding="utf-8")
+            f.create_dataset(
+                "goniometer/names", data=peaks_handler.goniometer.names_raw, dtype=dt
+            )
+
+        f.create_dataset(
+            "instrument/wavelength",
+            data=[peaks_handler.wavelength.min, peaks_handler.wavelength.max],
+        )
+        f.attrs["instrument"] = instrument
+
+    print(f"Saved {n_images} banks to {output_filename}")
+
+
+def run_merge_images(input_pattern: str, output_filename: str):
+    if " " in input_pattern:
+        h5_files = []
+        for p in input_pattern.split():
+            h5_files.extend(glob.glob(p))
+    else:
+        h5_files = glob.glob(input_pattern)
+
+    h5_files = sorted(list(set(h5_files)))
+
+    if not h5_files:
+        raise ValueError(f"No files found matching: {input_pattern}")
+
+    print(f"Found {len(h5_files)} files. Merging...")
+    merger = ImageStackMerger(h5_files)
+    merger.merge(output_filename)
+    print(f"Successfully created {output_filename}")
